changed package name and pattern strategy
the packge name is according to the github uri now.
note that i've changed my user name from daniloreinert to reinert only.
also i've changed the strategy pattern to tempate method.
it allows reuse a lot of common methods and it's more apropriated for
the complexty of the algorithm creation.