changes regarding Collections
I tried to create a Collection Map of all existing implementations of
Collection, but I figured out that it would be too hard, considering
others implementations like Google Collections. My itention is to
distinguish between Collection implementions and Collection wrappers
(used to build cusotmized array schemas). So I decided that any class
that implements AbstractCollection is an Collection Implementation,
otherwise (if only implements Collection or Iterable) is an Collection
Wrapper.