Many changes
Changed the build process of SchemaGenerators (from Factory to Builder
pattern - step-by-step way)

Changed the use of $schema attribute. It cant be set at annotation
anymore. It is put by the generator, depending on its version.

At the building process of the generator you can configure to put this
variable or not.